+++ demo/src/org/ode4j/democpp/DemoHeightfield.java
@@ -181,8 +181,8 @@ class DemoHeightfield extends dsFunctions {
 		dIASSERT( x < HFIELD_WSTEP );
 		dIASSERT( z < HFIELD_DSTEP );
 
-		double fx = ( ((double)x) - ( HFIELD_WSTEP-1 )/2 ) / (double)( HFIELD_WSTEP-1 );
-		double fz = ( ((double)z) - ( HFIELD_DSTEP-1 )/2 ) / (double)( HFIELD_DSTEP-1 );
+		double fx = ( ((double)x) - ( HFIELD_WSTEP-1 )/2 ) / ( HFIELD_WSTEP-1 );
+		double fz = ( ((double)z) - ( HFIELD_DSTEP-1 )/2 ) / ( HFIELD_DSTEP-1 );
 
 		// Create an interesting 'hump' shape
 		double h = ( 1.0 ) + ( ( -16.0 ) * ( fx*fx*fx + fz*fz*fz ) );
