+++ demo/src/org/ode4j/drawstuff/internal/DrawStuffNull.java
@@ -166,7 +166,7 @@ public class DrawStuffNull implements DrawStuffApi {
 			} else {
 				long timeUsed = 5000 + (startTime - System.currentTimeMillis());
 				startTime = System.currentTimeMillis() + 5000;
-				System.out.println(fps + " frames in " + (float) (timeUsed / 1000f) + " seconds = "
+				System.out.println(fps + " frames in " + (timeUsed / 1000f) + " seconds = "
 						+ (fps / (timeUsed / 1000f)));
 				fps = 0;
 //				loops++;
